+++ Source/Core/Core/HW/VideoInterface.cpp
@@ -50,7 +50,7 @@ static UVIBorderBlankRegister    m_BorderHBlank;
 // 0xcc002076 - 0xcc00207f is full of 0x00FF: unknown
 // 0xcc002080 - 0xcc002100 even more unknown
 
-u32 TargetRefreshRate = 0;
+float TargetRefreshRate = 0;
 
 static u32 TicksPerFrame = 0;
 static u32 s_lineCount = 0;
